% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allReviews.R
\name{allReviews}
\alias{allReviews}
\title{allReviews}
\usage{
allReviews(game_id = 0)
}
\arguments{
\item{game_id}{The game id. Integer or string.
Can be found at: https://steamdb.info/apps/}
}
\value{
A data.frame containing review and reviewer information
}
\description{
allReviews
}
\examples{
\dontrun{

posReviews(game_id = '976730')

}


}
