% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/appDetails.R
\name{appDetails}
\alias{appDetails}
\title{appDetails}
\usage{
appDetails(game_id = 0)
}
\arguments{
\item{game_id}{Integer, required. The game or app id to search for.}
}
\value{
data frame of app details
}
\description{
Returns information on any game or app on the Steam store.
}
\examples{

\dontrun{

appDetails(game_id = 50)

appDetails(game_id = 60)

}


}
