% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gameID.R
\name{gameID}
\alias{gameID}
\title{gameID}
\usage{
gameID(game_name = "")
}
\arguments{
\item{game_name}{String, required. Name of game.}
}
\value{
data.frame contraining possible matches.
}
\description{
gameID
}
\examples{

\dontrun{

gameID("Farming Simulator")

}

}
