% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gameList.R
\name{gameList}
\alias{gameList}
\title{gameList}
\usage{
gameList()
}
\value{
data.frame of all games on Steam along with their game ID
}
\description{
Returns a data.frame of all games from the Steam library with their appID.
See \url{http://api.steampowered.com/ISteamApps/GetAppList/v0002/?key=STEAMKEY&format=json} for more information.
}
\examples{
\dontrun{

gameList()

}


}
