% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getReviews.R
\name{getReviews}
\alias{getReviews}
\title{getReviews}
\usage{
getReviews(game_id = 0)
}
\arguments{
\item{game_id}{The game id. Can be found at: https://steamdb.info/apps/}
}
\value{
Character, required.
A dataframe containing review and reviewer information
}
\description{
getReviews
}
\examples{
\dontrun{

getReviews(game_id = '976730')

}


}
