% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpReviews.R
\name{helpReviews}
\alias{helpReviews}
\title{helpReviews}
\usage{
helpReviews(game_id = 0, day_range = 365)
}
\arguments{
\item{game_id}{The game id. Integer or string.
Can be found at: https://steamdb.info/apps/}

\item{day_range}{How many days to search for between 1 and 365.}
}
\value{
A data.frame containing review and reviewer information
}
\description{
helpReviews
}
\examples{
\dontrun{

posReviews(game_id = '976730')

}


}
