% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/negReviews.R
\name{negReviews}
\alias{negReviews}
\title{negReviews}
\usage{
negReviews(game_id = 0)
}
\arguments{
\item{game_id}{The game id. Integer or string.
Can be found at: https://steamdb.info/apps/}
}
\value{
A data.frame containing review and reviewer information
}
\description{
negReviews
}
\examples{
\dontrun{

getReviews(game_id = '976730')

}


}
