% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posReviews.R
\name{posReviews}
\alias{posReviews}
\title{posReviews}
\usage{
posReviews(game_id = 0)
}
\arguments{
\item{game_id}{The game id. Integer or string.
Can be found at: https://steamdb.info/apps/}
}
\value{
A data.frame containing review and reviewer information
}
\description{
posReviews
}
\examples{
\dontrun{

getReviews(game_id = '976730')

}


}
